+++ lib/SampleService/core/samples.py
@@ -154,6 +154,24 @@ class Samples:
                       _SampleAccessType.WRITE: 'cannot write to',
                       _SampleAccessType.READ: 'cannot read'}
 
+    def _check_batch_perms(
+        self,
+        ids_: List[UUID],
+        user: Optional[UserID],
+        access: _SampleAccessType,
+        acls: List[SampleACL] = None,
+        as_admin: bool = False):
+            if as_admin:
+                return
+            if not acls:
+                acls = self._storage.get_sample_set_acls(ids_)
+            levels = [self._get_access_level(acl, user) for acl in acls]
+            for i, level in enumerate(levels):
+                if level < access:
+                    uerr = f'User {user}' if user else 'Anonymous users'
+                    errmsg = f'{uerr} {self._unauth_errmap[access]} sample {ids_[i]}'
+                    raise _UnauthorizedError(errmsg)
+
     def _get_access_level(self, acls: SampleACL, user: Optional[UserID]):
         if user == acls.owner:
             return _SampleAccessType.OWNER
@@ -440,6 +458,37 @@ class Samples:
         wsids = None if as_admin else self._ws.get_user_workspaces(user)
         # TODO DATALINK what about deleted objects? Currently not handled
         return self._storage.get_links_from_sample(sample, wsids, timestamp), timestamp
+
+    def get_batch_links_from_sample_set(
+            self,
+            user: Optional[UserID],
+            samples: List[SampleAddress],
+            timestamp: datetime.datetime = None,
+            as_admin: bool = False) -> Tuple[List[DataLink], datetime.datetime]:
+        '''
+        A batch version of get_links_from_sample. Gets a set of  data links originating
+        from multiple samples in a given sampleset at a particular time.
+
+        :param user: the user requesting the links or None if the user is anonymous.
+        :param samples: the list of sample ids from which the links originate.
+        :param timestamp: the timestamp during which the links should be active, defaulting to
+            the current time.
+        :param as_admin: allow link retrieval to proceed if user does not have
+            appropriate permissions.
+        :returns: a tuple consisting of a list of links and the timestamp used to query the links.
+        :raises UnauthorizedError: if the user does not have read permission for the sample.
+        :raises NoSuchSampleError: if the sample does not exist.
+        :raises NoSuchSampleVersionError: if the sample version does not exist.
+        :raises NoSuchUserError: if the user does not exist.
+        '''
+        _not_falsy(samples, 'samples')
+        timestamp = self._resolve_timestamp(timestamp)
+        wsids = None if as_admin else self._ws.get_user_workspaces(user)
+        # checks for all sample acls in one query
+        sampleids = [s.sampleid for s in samples]
+        self._check_batch_perms(sampleids, user, _SampleAccessType.READ, as_admin=as_admin)
+        return_links = self._storage.get_batch_links_from_samples(samples, wsids, timestamp)
+        return return_links, timestamp
 
     def _resolve_timestamp(self, timestamp: datetime.datetime = None) -> datetime.datetime:
         if timestamp: